修复安装工具bug

+++ app/src/main/assets/bin/include/core/del_rootfs.sh
@@ -4,10 +4,15 @@
 # license: GPL-v2.0
 #
 del_rootfs() {
-echo "- 正在关闭"
-stop_rootfs
-umount_part
-echo "- 正在删除"
-rm -rf $rootfs/*
-echo "- 删除完成"
+if [ -d "$rootfs/proc/" ];then
+  echo "- 正在关闭"
+  stop_rootfs
+  umount_part
+  echo "- 正在删除"
+  rm -rf $rootfs/*
+  echo "- 删除完成"
+  else
+  echo "- 警告 "
+  echo "- 找不到Rootfs路径"
+fi
 }
